+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### September
 
+#### 22 September 2018
+
+* Add sprite ID validation to animated sprites.
+
 #### 21 September 2018
 
 * Rename 'doc' directory to 'docs'.